Refactor case loading and session management into a utility function
