+++ src/main/functions/util/commonUtil.ts
@@ -1,5 +1,7 @@
+import type { Request } from 'express';
+import type { LoggerInstance } from 'winston';
+
 import { FinremCaseData } from 'app/case/definition';
-import { LoggerInstance } from 'winston';
 
 import { getSystemUser } from '../../app/auth/user';
 import { getCaseApi } from '../../app/case/case-api';
@@ -28,5 +30,39 @@ export async function getHomePageForUser(userDetails: UserDetails): Promise<User
   } else {
     logger.info('Routing to : ', RouteNames.enterCaseNumber);
     return { url: RouteNames.enterCaseNumber };
+  }
+}
+
+/**
+ * Loads a case by reference from CCD and stores it in the session
+ * @param req - Express request object with session
+ * @param caseReference - Case reference/ID (with or without hyphens)
+ * @param logger - Logger instance
+ * @returns Promise<FinremCaseData> - The loaded case data
+ * @throws Error if case is not found or cannot be loaded
+ */
+export async function loadCaseAndReloadSession(
+  req: Request,
+  caseReference: string,
+  logger: LoggerInstance
+): Promise<FinremCaseData> {
+  const ccdUrl = require('config').get('services.case.url');
+  const caseId = caseReference.replace(/-/g, '');
+  
+  logger.info(`Loading case ${caseId} from CCD backend: ${ccdUrl}`);
+  
+  try {
+    const systemUser = await getSystemUser();
+    const caseApi = getCaseApi(systemUser, logger);
+    const caseData = await caseApi.getCaseById(caseId);
+    
+    logger.info(`Case ${caseId} successfully loaded from CCD`);
+    
+    req.session.caseData = caseData;
+    
+    return caseData;
+  } catch (error) {
+    logger.error(`Failed to load case ${caseId} from CCD:`, error);
+    throw error;
   }
 }
